+++ backend/prisma/migrations/20251028185710_add_payment_order_relation/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "public"."ExternalProductionOrder" ADD COLUMN     "assemblerPaymentId" TEXT;
+
+-- AddForeignKey
+ALTER TABLE "public"."ExternalProductionOrder" ADD CONSTRAINT "ExternalProductionOrder_assemblerPaymentId_fkey" FOREIGN KEY ("assemblerPaymentId") REFERENCES "public"."AssemblerPayment"("id") ON DELETE SET NULL ON UPDATE CASCADE;